fix(build): resolver fallos de CI en build.yml y CMakeLists
- CMakeLists.txt: quitar referencia a spatial/cue_based_spatial.cpp
  (nunca existió como .cpp, la clase CueBasedSpatial es header-only).
  Esto tumbaba el configure de CMake antes de compilar nada.
- build.yml: corregir typo env.ANDROID_NDK_VERSION -> env.ANDROID_NDK_HOME
  en el job build-magisk (variable inexistente dejaba NDK_HOME vacío).
- dsp/Compressor.cpp: renombrar envelope_ -> env_ para matchear el
  miembro real declarado en include/Compressor.h.
- dsp/ParametricEQ.cpp: eliminar tipo BiquadCoeff fantasma y setCoeff()
  inexistente; escribir coeficientes directo en la struct Biquad real
  del header. Renombrar process()->processSample() por banda y agregar
  noexcept faltante en process() para matchear la firma del header.
- neuromorphic/volterra_h2_symmetric.cpp: reescribir constructor sin
  try/catch/throw (el proyecto compila con -fno-exceptions). Ahora es
  nothrow-safe: si falla una allocación, libera lo reservado y deja
  m_kernels_ready=false en vez de lanzar std::bad_alloc.
- neuromorphic/volterra_h2_symmetric.hpp: agregar getter público
  isReady() para exponer m_kernels_ready.
- neuromorphic/neuro_cochlear_manifold.cpp: quitar try/catch redundante
  ya incompatible con -fno-exceptions; verificar por punteros +
  volterra->isReady().
- anti_dolby.cpp: std::expf (no existe en el standard) -> std::exp.

Validado localmente: configure de CMake limpio (sin el error de source
file faltante), cross-compile syntax check con aarch64-linux-gnu-g++
sin errores en los archivos tocados, y suite completa de tests
(app/src/main/cpp/tests) compila y pasa 4/4 con host g++.

regla de oro: no borramos, solo mejoramos, reparamos, optimizamos.

+++ app/src/main/cpp/anti_dolby.cpp
@@ -48,7 +48,7 @@ void AntiDolbyState::tick(float dt) {
         std::lock_guard<std::mutex> lk(mtx);
         // Coeficientes RC -> alpha = 1 - exp(-dt / tau)
         float tau = (targetWidener < smoothedWidener) ? attackTau : releaseTau;
-        alpha = 1.0f - std::expf(-dt / tau);
+        alpha = 1.0f - std::exp(-dt / tau);
         smoothedWidener = smoothedWidener + alpha * (targetWidener - smoothedWidener);
     }
     // Publicar el valor suavizado (no espaciado de memory_order estricto requerido)
